Register views of the `peering` app
