+++ templates/peering/internetexchangepeeringsession/_base.html
@@ -20,14 +20,14 @@
 {% endblock %}
 {% block breadcrumb %}
 <li class="breadcrumb-item"><a href="{% url 'peering:internetexchange_list' %}">Internet Exchanges</a></li>
-<li class="breadcrumb-item"><a href="{% url 'peering:internetexchange_view' pk=instance.ixp_connection.internet_exchange_point.pk %}">{{ instance.ixp_connection.internet_exchange_point }}</a></li>
+<li class="breadcrumb-item"><a href="{% url 'peering:internetexchange' pk=instance.ixp_connection.internet_exchange_point.pk %}">{{ instance.ixp_connection.internet_exchange_point }}</a></li>
 <li class="breadcrumb-item"><a href="{% url 'peering:internetexchange_peering_sessions' pk=instance.ixp_connection.internet_exchange_point.pk %}">Peering Sessions</a></li>
 <li class="breadcrumb-item active" aria-current="page">{% block title %}{{ instance }}{% endblock %}</li>
 {% endblock %}
 {% block content %}
 <ul class="nav nav-tabs mb-3">
   <li class="nav-item">
-    <a class="nav-link{% if tab == 'main' %} active{% endif %}" href="{% url 'peering:internetexchangepeeringsession_view' pk=instance.pk %}">
+    <a class="nav-link{% if tab == 'main' %} active{% endif %}" href="{% url 'peering:internetexchangepeeringsession' pk=instance.pk %}">
       <i class="fa-fw fa-solid fa-info-circle"></i> Info
     </a>
   </li>
